+++ tests/test.py
@@ -38,19 +38,19 @@ if __name__ == "__main__":
     # Load the dataset
     if dataset == 2:
         data, freq, fc_hor, mis_val, eq_len = convert_tsf_to_dataframe(paths[2], 0)
-        d = np.array([data.loc[i,"series_value"].to_numpy() for i in range(data.shape[0])], order='C').T
+        d = np.array([data.loc[i,"series_value"].to_numpy() for i in range(data.shape[0])], order='C', dtype= np.float32).T
     elif dataset == 4:
         data = pd.read_csv(paths[dataset])
         data = data.drop(['Time','Unix', 'Issues'],axis=1)
-        d = np.ascontiguousarray(data.to_numpy(dtype=np.float64))
+        d = np.ascontiguousarray(data.to_numpy(), dtype=np.float32)
         d = d[:100000,:]
     elif dataset == 3:
         data = pd.read_csv(paths[dataset])
-        d = np.ascontiguousarray(data.to_numpy(dtype=np.float64))
+        d = np.ascontiguousarray(data.to_numpy(), dtype=np.float32)
     else:
         data = pd.read_csv(paths[dataset], sep=r'\s+')
         data = data.drop(data.columns[[0]], axis=1)
-        d = np.ascontiguousarray(data.to_numpy())
+        d = np.ascontiguousarray(data.to_numpy(), dtype=np.float32)
     
 
     r = 8#find_width_discr(d, window_size, K)
